refactor: put braintree mappers in separate namespace
