+++ src/Processor/Braintree/Mapper/PaymentMethodMapper.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace TeamGantt\Dues\Processor\Braintree;
+namespace TeamGantt\Dues\Processor\Braintree\Mapper;
 
 use InvalidArgumentException;
 use TeamGantt\Dues\Arr;
